+++ ddtrace/contrib/internal/pyramid/constants.py
@@ -1,0 +1,6 @@
+SETTINGS_SERVICE = "datadog_trace_service"
+SETTINGS_TRACER = "datadog_tracer"
+SETTINGS_TRACE_ENABLED = "datadog_trace_enabled"
+SETTINGS_DISTRIBUTED_TRACING = "datadog_distributed_tracing"
+SETTINGS_ANALYTICS_ENABLED = "datadog_analytics_enabled"
+SETTINGS_ANALYTICS_SAMPLE_RATE = "datadog_analytics_sample_rate"
