+++ ddtrace/contrib/internal/requests/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_SERVICE = "requests"
