+++ ddtrace/contrib/pyodbc/__init__.py
@@ -60,7 +60,11 @@ required_modules = ["pyodbc"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.pyodbc.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.pyodbc.patch import get_version
+        from ..internal.pyodbc.patch import patch
 
         __all__ = ["patch", "get_version"]
