+++ ddtrace/contrib/pyramid/constants.py
@@ -1,6 +1,15 @@
-SETTINGS_SERVICE = "datadog_trace_service"
-SETTINGS_TRACER = "datadog_tracer"
-SETTINGS_TRACE_ENABLED = "datadog_trace_enabled"
-SETTINGS_DISTRIBUTED_TRACING = "datadog_distributed_tracing"
-SETTINGS_ANALYTICS_ENABLED = "datadog_analytics_enabled"
-SETTINGS_ANALYTICS_SAMPLE_RATE = "datadog_analytics_sample_rate"
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
+from ddtrace.vendor.debtcollector import deprecate
+
+from ..internal.pyramid.constants import *  # noqa: F401,F403
+
+
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
+
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
