+++ ddtrace/contrib/redis/__init__.py
@@ -74,7 +74,11 @@ required_modules = ["redis", "redis.client"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.redis.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.redis.patch import get_version
+        from ..internal.redis.patch import patch
 
         __all__ = ["patch", "get_version"]
