+++ ddtrace/contrib/rediscluster/__init__.py
@@ -55,7 +55,11 @@ required_modules = ["rediscluster", "rediscluster.client"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.rediscluster.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.rediscluster.patch import get_version
+        from ..internal.rediscluster.patch import patch
 
         __all__ = ["patch", "get_version"]
