+++ ddtrace/contrib/requests/__init__.py
@@ -79,9 +79,13 @@ required_modules = ["requests"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
-        from .session import TracedSession
+        # Required to allow users to import from `ddtrace.contrib.requests.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.requests.patch import get_version
+        from ..internal.requests.patch import patch
+        from ..internal.requests.patch import unpatch
+        from ..internal.requests.session import TracedSession
 
         __all__ = ["patch", "unpatch", "TracedSession", "get_version"]
