+++ tests/.suitespec.json
@@ -145,7 +145,9 @@
         ],
         "redis": [
             "ddtrace/contrib/rediscluster/*",
+            "ddtrace/contrib/internal/rediscluster/*",
             "ddtrace/contrib/redis/*",
+            "ddtrace/contrib/internal/redis/*",
             "ddtrace/contrib/aredis/*",
             "ddtrace/contrib/internal/aredis/*",
             "ddtrace/contrib/yaaredis/*",
@@ -209,10 +211,12 @@
             "ddtrace/contrib/tornado/*"
         ],
         "pynamodb": [
-            "ddtrace/contrib/pynamodb/*"
+            "ddtrace/contrib/pynamodb/*",
+            "ddtrace/contrib/internal/pynamodb/*"
         ],
         "pyodbc": [
-            "ddtrace/contrib/pyodbc/*"
+            "ddtrace/contrib/pyodbc/*",
+            "ddtrace/contrib/internal/pyodbc/*"
         ],
         "flask": [
             "ddtrace/contrib/flask/*",
@@ -253,10 +257,12 @@
             "ddtrace/contrib/sanic/*"
         ],
         "requests": [
-            "ddtrace/contrib/requests/*"
+            "ddtrace/contrib/requests/*",
+            "ddtrace/contrib/internal/requests/*"
         ],
         "pyramid": [
-            "ddtrace/contrib/pyramid/*"
+            "ddtrace/contrib/pyramid/*",
+            "ddtrace/contrib/internal/pyramid/*"
         ],
         "langchain": [
             "ddtrace/contrib/langchain/*",
